requirements.txt
pika
python-socketio
eventlet